Added a working display amount shortener
If there are 1000 or more of a single item, the item list will now show the item count to 3 significant figures (rounded down), while the item focus card still shows the exact amount. This prevents the deliberate erroring out if there are more than 100000 of the same item present in the system, as well as looking rather nice (in the author's opinion).