Update mssInstaller.lua
Covers checking that mssInstaller.lua is being ran on a Crafty Turtle, is connected to a Wired Modem, that there is a network-attached common code Disk Drive and Floppy Disk, and that there is a network-attached manifest file Disk Drive and Floppy Disk. The next step would be defining the client import buffer inventory.